# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handlers.AnalyticsPopularStationDTO:
    properties:
      country:
        example: USA
        type: string
      favicon:
        example: https://cdn.example.com/favicon.png
        type: string
      name:
        example: Rock FM
        type: string
      plays:
        example: 1520
        type: integer
      station_id:
        example: abc123
        type: string
      url:
        example: https://stream.example.com/live
        type: string
    type: object
...
  internal_handlers.CountData:
    properties:
      count:
        example: 1234
        type: integer
    type: object
  internal_handlers.CountResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handlers.CountData'
      success:
        example: true
        type: boolean
    type: object
...
  internal_handlers.ErrorDetail:
    properties:
      code:
        example: STATION_NOT_FOUND
        type: string
      field:
        example: email (only present on validation errors)
        type: string
      message:
        example: station not found
        type: string
    type: object
  internal_handlers.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_handlers.ErrorDetail'
    type: object
...
  internal_handlers.GuestDetailsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/radio-backend_internal_domain.GuestDetail'
        type: array
      success:
        example: true
        type: boolean
    type: object
...
  internal_handlers.PopularStationsAnalyticsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_handlers.AnalyticsPopularStationDTO'
        type: array
      success:
        example: true
        type: boolean
    type: object
...
  internal_handlers.SimpleErrorResponse:
    properties:
      error:
        example: authentication required
        type: string
    type: object
...
  internal_handlers.StreamStatsResponse:
    properties:
      active_streams:
        type: integer
      streams:
        items:
          $ref: '#/definitions/radio-backend_internal_services.StreamInfo'
        type: array
    type: object
...
  internal_handlers.TrendingSearchDTO:
    properties:
      count:
        example: 456
        type: integer
      percentage:
        example: 12.5
        type: number
      search_term:
        example: rock
        type: string
    type: object
  internal_handlers.TrendingSearchesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_handlers.TrendingSearchDTO'
        type: array
      success:
        example: true
        type: boolean
    type: object
...
  radio-backend_internal_domain.GuestDetail:
    properties:
      endpoints:
        items:
          $ref: '#/definitions/radio-backend_internal_domain.GuestEndpointStat'
        type: array
      first_seen:
        type: string
      ip_address:
        type: string
      last_seen:
        type: string
      total_requests:
        type: integer
      unique_endpoints:
        type: integer
      user_agent:
        type: string
    type: object
  radio-backend_internal_domain.GuestEndpointStat:
    properties:
      count:
        type: integer
      method:
        type: string
      path:
        type: string
    type: object
...
  radio-backend_internal_services.StreamInfo:
    properties:
      bitrateKbps:
        format: float64
        type: number
      bytesStreamed:
        format: int64
        type: integer
      duration:
        $ref: '#/definitions/time.Duration'
      lastUpdate:
        type: string
      sessionID:
        type: string
    type: object
...
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    format: int64
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
host: localhost:8080
info:
  contact:
    email: support@radiobackend.com
    name: API Support
  description: "Radio streaming API with JWT authentication, analytics and multi-language
    support (i18n).\n\n**Security Features (v2.1):**\n- \U0001F510 Timing attack prevention
    (constant-time operations)\n- \U0001F512 HTTPS enforcement (production)\n- \U0001F6AB
    Account lockout (10 attempts = 30 min lockout)\n- \U0001F4E7 Email rate limiting
    (10 attempts/hour per email)\n- \U0001F511 Password strength enforcement (special
    chars + 47 common passwords blocked)\n- \U0001F6E1️ Session hijacking protection
    (User-Agent validation)\n- \U0001F4DD Security event logging (failed attempts
    with metadata)\n- ⚡ Rate limiting (5 attempts/15min per IP)\n- \U0001F510 JWT
    with RS256 + token revocation\n- \U0001F916 Bot protection via shared secret (X-Rradio-Secret
    header)\n\n**Internationalization:**\nSupported languages: Spanish (es), English
    (en), French (fr), German (de).\nUse the 'lang' parameter or 'Accept-Language'
    header to specify the desired language.\n\n**Security Score:** 100/100 (audited
    Dec 2025)"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Radio Backend API
  version: "2.1"
paths:
...
  /analytics/searches/trending:
    get:
      consumes:
      - application/json
      description: Returns the most frequent search terms in a specific time period
        ordered by frequency. Includes absolute count and percentage of total.
      parameters:
      - default: day
        description: 'Time range: hour, day, week, month'
        enum:
        - hour
        - day
        - week
        - month
        in: query
        name: range
        type: string
      - default: 10
        description: Maximum number of results to return
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Trending searches statistics
          schema:
            $ref: '#/definitions/internal_handlers.TrendingSearchesResponse'
        "401":
          description: Invalid or missing authentication token
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Access denied - Admin users only
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Trending searches
      tags:
      - Analytics
  /analytics/stations/popular:
    get:
      consumes:
      - application/json
      description: Returns the most played stations in a specific time period ordered
        by number of plays. Includes complete information for each station.
      parameters:
      - default: day
        description: 'Time range: hour, day, week, month'
        enum:
        - hour
        - day
        - week
        - month
        in: query
        name: range
        type: string
      - default: 10
        description: Maximum number of results to return
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Popular stations statistics
          schema:
            $ref: '#/definitions/internal_handlers.PopularStationsAnalyticsResponse'
        "401":
          description: Invalid or missing authentication token
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Access denied - Admin users only
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Popular stations statistics
      tags:
      - Analytics
  /analytics/users/active:
    get:
      consumes:
      - application/json
      description: Returns the number of authenticated active users in the last 24
        hours. Active users are those who have made at least one authenticated request
        in the specified period.
      produces:
      - application/json
      responses:
        "200":
          description: Active users count
          schema:
            $ref: '#/definitions/internal_handlers.CountResponse'
        "401":
          description: Invalid or missing authentication token
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Access denied - Admin users only
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Active users
      tags:
      - Analytics
  /analytics/users/guest:
    get:
      consumes:
      - application/json
      description: Returns the number of guest (unauthenticated) active users in the
        last 24 hours. Guest users are identified by their unique IP address and represent
        users using the application without registering.
      produces:
      - application/json
      responses:
        "200":
          description: Guest users count
          schema:
            $ref: '#/definitions/internal_handlers.CountResponse'
        "401":
          description: Invalid or missing authentication token
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Access denied - Admin users only
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Active guest users
      tags:
      - Analytics
  /analytics/users/guest/details:
    get:
      consumes:
      - application/json
      description: Returns request details for each guest (unauthenticated) IP address
        in the given time range, ordered by total requests descending. Useful for
        understanding guest behavior patterns.
      parameters:
      - default: day
        description: 'Time range: hour, day, week, month'
        enum:
        - hour
        - day
        - week
        - month
        in: query
        name: range
        type: string
      - default: 50
        description: Maximum number of IPs to return
        in: query
        maximum: 500
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Guest details list
          schema:
            $ref: '#/definitions/internal_handlers.GuestDetailsResponse'
        "401":
          description: Invalid or missing authentication token
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Access denied - Admin users only
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Guest users details
      tags:
      - Analytics
...
  /stream/proxy:
    get:
      description: Hace proxy del stream de audio de una estación validando el token
        de sesión
      parameters:
      - description: Token JWT de la sesión
        in: query
        name: token
        required: true
        type: string
      produces:
      - audio/mpeg
      responses:
        "200":
          description: Audio stream (audio/mpeg)
          schema:
            type: file
        "401":
          description: Missing, invalid or expired token
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Station not found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "502":
          description: Failed to connect to the upstream stream
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      summary: Proxy de stream de audio
      tags:
      - Streaming
...
  /stream/stats:
    get:
      description: Retorna información sobre streams actualmente activos (solo admin)
      produces:
      - application/json
      responses:
        "200":
          description: Active stream statistics
          schema:
            $ref: '#/definitions/internal_handlers.StreamStatsResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener estadísticas de streams activos
      tags:
      - Streaming
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  SharedSecret:
    description: Shared secret key required for public station and SEO endpoints (bot
      protection).
    in: header
    name: X-Rradio-Secret
    type: apiKey
swagger: "2.0"